Fix some package and javadoc
